articles/generative_ai/diffusion_models_fundamentals/pyproject.toml
[tool.poetry]
name = "diffusion-models-fundamentals"
version = "0.1.0"
description = "Diffusion model training"
authors = ["Anca <ancamuscalagiu1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyarrow = "14.0.1"
torch = "2.3.1"
torchaudio = "2.3.1"
gcsfs = "2024.3.1"
accelerate = "0.33.0"
torchvision = "0.18.1"
transformers = "4.44.0"
pytorch-fid = "0.3.0"
datasets = "2.19.2"
diffusers = "0.30.0"
wandb = "^0.18.3"
pyyaml = "^6.0.2"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
